+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 
+- Avoid using the deprecated `$GLOBALS['TSFE']->fe_user` (#968)
 - Drop obsolete entries from the `ext_emconf.php` (#860)
 
 ## 6.4.0